Put far side switch strategies in folder, added far side scale strategies folder as well
